Backend/encryption/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pycryptodome==3.20.0
python-dotenv==1.0.0
requests==2.31.0
stegano==0.11.2
Pillow==10.1.0
